Add files via upload
This assignment was locked Oct 7, 2019 at 11:55pm.
Write a Python program for optimizing the SVM hinge loss using
gradient descent algorithm. The input and output should be the same as for
assignment 2.

Test your program with the data file testSVM.data and training label file testSVM.trainlabels

Convert label 0 to -1 so that labels r_i's are either +1 or -1. This is
necessary for the gradient descent to work.

Use eta=.001 and stopping condition of while(abs(prevError - error) > .000000001).
Note the absolute value to account for instability in the gradient for hinge
loss. The converged solution with the hinge loss for the test data given above would be approximately

w = (1.460557, -0.459554)
w0 = -2.002468
Dist to origin= 1.307820